feat: Update Bike ride data and location tracking
This commit updates the Bike application to use floating-point values for distance measurements and improves location tracking.

Key changes:

- Updated distance-related fields (`currentTripDistance`, `totalDistance`, `remainingDistance`) in `BikeRideInfo` and related composables to use `Float` for consistency.
- Refactored `UnifiedLocationRepositoryImpl` to correctly calculate and expose `traveledDistanceFlow` and `remainingDistanceFlow` using location updates.
- The demo mode for `BikeViewModel` is disabled to reflect real location updates.
- Updated example data and previews to use `Float` for distance values.

+++ applications/ashbike/src/main/java/com/ylabz/basepro/applications/bike/ui/components/home/main/SpeedAndProgressCard.kt
@@ -39,8 +39,8 @@ import com.ylabz.basepro.feature.weather.ui.components.combine.WindDirectionDial
 fun SpeedAndProgressCard(
     modifier: Modifier = Modifier.fillMaxSize(),
     currentSpeed: Double,
-    currentTripDistance: Double,
-    totalDistance: Double,
+    currentTripDistance: Float,
+    totalDistance: Float,
     windDegree: Float,
     windSpeed: Float,
     weatherConditionText: String,
@@ -144,8 +144,8 @@ fun SpeedAndProgressCard(
 fun SpeedAndProgressCardPreview() {
     SpeedAndProgressCard(
         currentSpeed = 25.5,
-        currentTripDistance = 10.0,
-        totalDistance = 50.0,
+        currentTripDistance = 10.0f,
+        totalDistance = 50.0f,
         windDegree = 120f,
         windSpeed = 5.0f,
         weatherConditionText = WeatherConditionUnif.RAINY.name,
